Inline unused type
